feature: beaver integration - #20
