+++ tests/cypress/support/commands.ts
@@ -1,15 +1,16 @@
-import { FormSubmitionSettings, submitForm } from "./commands/submit-form";
+import { FormSubmitionSettings, submitForm } from "./commands/submitForm";
 import Login from "./commands/login";
-import SafeType from "./commands/safe-type";
+import SafeType from "./commands/safeType";
 import {
 	RemovePosts,
 	Settings as RemovePostSettings,
-} from "./commands/remove-posts";
+} from "./commands/removePosts";
 import {
 	RemoveUsers,
 	Settings as RemoveUsersSettings,
-} from "./commands/remove-users";
+} from "./commands/removeUsers";
 import Chainable = Cypress.Chainable;
+import { assertProcaptchaExistence } from "@support/commands/assertProcaptchaExistence";
 
 declare global {
 	namespace Cypress {
@@ -34,71 +35,12 @@ declare global {
 				selectorOrElement: string | JQuery,
 			): Chainable<JQuery>;
 
-			togglePlugins(
-				enable: boolean,
-				pluginSlugs: string[],
-			): Chainable<string[]>;
+			test();
 		}
 	}
 }
 
-Cypress.Commands.add(
-	"togglePlugins",
-	(enable: boolean, pluginSlugs: string[]) => {
-		cy.login();
-
-		const affectedPluginSlugs: string[] = [];
-		const action = enable ? "activate" : "deactivate";
-
-		return cy
-			.wrap(pluginSlugs)
-			.each((slug: string) => {
-				let selector = "#" + action + "-" + slug;
-
-				cy.visit("/wp-admin/plugins.php");
-
-				cy.get("body").then(($body) => {
-					// optional, as the plugin may be already active (avoid breaks if tests are run locally).
-					if (0 === $body.find(selector).length) {
-						return;
-					}
-
-					// visit instead of the click, as some plugins have deactivation survey popups.
-					cy.get(selector)
-						.invoke("attr", "href")
-						.then((href) => {
-							cy.visit("/wp-admin/" + href);
-						});
-
-					// check for url instead of the notice, as some plugins (like BBPress) make a redirect.
-					cy.url().should("not.equal", "/wp-admin/plugins.php");
-
-					affectedPluginSlugs.push(slug);
-				});
-			})
-			.then(() => affectedPluginSlugs);
-	},
-);
-
-Cypress.Commands.add(
-	"assertProcaptchaExistence",
-	(shouldExist: boolean, selectorOrElement: string | JQuery): void => {
-		const element: Chainable<JQuery> =
-			"string" === typeof selectorOrElement
-				? cy.get(<string>selectorOrElement)
-				: cy.wrap(<JQuery>selectorOrElement);
-
-		const assertion = shouldExist ? "exist" : "not.exist";
-
-		element.find(".prosopo-procaptcha-wp-widget").should(assertion);
-
-		// check only in the 'exist' case, as for "not.exist" case,
-		// the script could have been added for the other form.
-		if (shouldExist) {
-			cy.get("script#prosopo-procaptcha-js").should("exist");
-		}
-	},
-);
+Cypress.Commands.add("assertProcaptchaExistence", assertProcaptchaExistence);
 
 // with workarounds, otherwise .type below writes too fast, or just fails.
 Cypress.Commands.add(
