+++ tests/cypress/support/commands/removePosts.ts
