+++ tests/cypress/support/commands/removeUsers.ts
