+++ tests/cypress/support/commands/safeType.ts
